pyproject.toml
[tool.pdm]
version = {source = "file", path = "common/__init__.py"}

[tool.pdm.dev-dependencies]
dev = ["pre-commit>=3.2.2"]

[project]
authors = [{name = "Murilo Cunha", email = "murilo@dataroots.io"}]
dependencies = [
    "pydantic>=1.10.7",
    "fastapi>=0.95.1",
    "modal-client>=0.49.2059",
]
description = ""
dynamic = ["version"]
license = {text = "MIT"}
name = "common"
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
modal-app = [
    "transformers~=4.28.1",
    "safetensors~=0.3.0",
    "accelerate~=0.18.0",
    "bitsandbytes~=0.38.1",
    "sentencepiece~=0.1.98",
    "hf-transfer~=0.1.3",
    "msgspec~=0.14.2",
]
[build-system]
build-backend = "pdm.backend"
requires      = ["pdm-backend"]
